Extract shared MarkdownViewer component and add editor shortcut (#485)
* feat: convert /skills command from full page to dialog overlay

Closes #470

* feat: render skill instructions as styled markdown in detail view

Uses the <markdown> element with SyntaxStyle, matching the pentest
report viewer pattern for consistent rich markdown rendering.

* style: fix lint and format issues in skills dialog

* style: fix README.md prettier formatting

* refactor: extract shared MarkdownViewer component and add [E] editor shortcut

- Create reusable MarkdownViewer component with themed syntax highlighting,
  scrollable content, and configurable header/footer/metadata slots
- Extract openFileInDefaultApp utility from open-report.ts for generic use
- Refactor ReportViewerDialog to use MarkdownViewer (no behavior change)
- Refactor skills detail view to use MarkdownViewer
- Add [E] keyboard shortcut in skills detail to open SKILL.md in editor

* fix: add renderNode callback to MarkdownViewer for theme-aware code blocks

CodeRenderable in opentui defaults to white text when tree-sitter can't
highlight (e.g., code blocks with no language tag). This was fixed in the
report-viewer-dialog via a renderNode callback (PR #473), but that fix
was lost when MarkdownViewer was extracted as a shared component.

Moves the renderNode callback into MarkdownViewer itself so all consumers
(report viewer, skills detail) get theme-aware code block colors.

---------
