+++ src/tui/components/report-viewer-dialog.tsx
@@ -1,14 +1,8 @@
-import { useRef, useMemo, useCallback } from "react";
+import { useMemo } from "react";
 import { useKeyboard } from "@opentui/react";
 import { useDimensions } from "../context/dimensions";
-import {
-  Renderable,
-  ScrollBoxRenderable,
-  SyntaxStyle,
-  type CodeRenderable,
-} from "@opentui/core";
-import type { Token } from "marked";
 import { useTheme } from "../theme";
+import { MarkdownViewer } from "./shared/markdown-viewer";
 import { DialogControls } from "./shared/dialog-controls";
 
 interface ReportViewerDialogProps {
@@ -26,48 +20,6 @@ export default function ReportViewerDialog({
 }: ReportViewerDialogProps) {
   const { colors } = useTheme();
   const dimensions = useDimensions();
-  const scrollRef = useRef<ScrollBoxRenderable>(null);
-
-  const syntaxStyle = useMemo(
-    () =>
-      SyntaxStyle.fromStyles({
-        default: { fg: colors.text },
-        "markup.heading": { fg: colors.markdownHeading, bold: true },
-        "markup.strong": { fg: colors.markdownStrong, bold: true },
-        "markup.italic": { fg: colors.markdownEmph, italic: true },
-        "markup.raw": { fg: colors.markdownCode },
-        "markup.raw.block": { fg: colors.markdownCode },
-        "markup.link": { fg: colors.markdownLink },
-        "markup.link.url": { fg: colors.markdownLink, dim: true },
-        "markup.link.label": { fg: colors.markdownLink, underline: true },
-        "markup.strikethrough": { fg: colors.textMuted, dim: true },
-        "markup.list": { fg: colors.primary },
-        "markup.quote": { fg: colors.textMuted, italic: true },
-        "punctuation.special": { fg: colors.border },
-      }),
-    [colors],
-  );
-
-  // Override code block rendering to set theme-aware fg color.
-  // Without this, CodeRenderable defaults to white text (opentui default),
-  // which is invisible on light backgrounds.
-  const renderNode = useCallback(
-    (
-      token: Token,
-      ctx: { defaultRender: () => Renderable | null },
-    ): Renderable | undefined | null => {
-      if (token.type === "code") {
-        const renderable = ctx.defaultRender();
-        if (renderable) {
-          (renderable as CodeRenderable).fg = colors.markdownCode;
-        }
-        return renderable;
-      }
-      return undefined;
-    },
-    [colors.markdownCode],
-  );
-
   const lineCount = useMemo(() => content.split("\n").length, [content]);
 
   useKeyboard((evt) => {
@@ -110,83 +62,28 @@ export default function ReportViewerDialog({
         borderStyle="single"
         flexDirection="column"
       >
-        {/* Header */}
-        <box
-          width="100%"
-          padding={1}
-          flexDirection="row"
-          justifyContent="space-between"
-        >
-          <text fg={colors.primary}>Pentest Report</text>
-          <text fg={colors.textMuted}>{reportPath}</text>
-        </box>
-
-        {/* Separator */}
-        <box width="100%" height={1}>
-          <text fg={colors.border}>{"─".repeat(panelWidth - 2)}</text>
-        </box>
-
-        {/* Report Content */}
-        <scrollbox
-          ref={scrollRef}
-          style={{
-            rootOptions: {
-              flexGrow: 1,
-              width: "100%",
-              overflow: "hidden",
-            },
-            contentOptions: {
-              paddingLeft: 2,
-              paddingRight: 2,
-              paddingTop: 1,
-              paddingBottom: 1,
-              flexDirection: "column",
-            },
-            scrollbarOptions: {
-              trackOptions: {
-                foregroundColor: colors.primary,
-                backgroundColor: colors.backgroundElement,
-              },
-            },
-          }}
-          stickyScroll={false}
-          focused={true}
-        >
-          <markdown
-            content={content}
-            syntaxStyle={syntaxStyle}
-            conceal={true}
-            renderNode={renderNode}
-          />
-        </scrollbox>
-
-        {/* Separator */}
-        <box width="100%" height={1}>
-          <text fg={colors.border}>{"─".repeat(panelWidth - 2)}</text>
-        </box>
-
-        {/* Footer */}
-        <box
-          width="100%"
-          padding={1}
-          flexDirection="row"
-          justifyContent="space-between"
-        >
-          <DialogControls
-            controls={[
-              ...(onOpenExternal
-                ? [
-                    {
-                      key: "E",
-                      label: "Open in Editor",
-                      variant: "primary" as const,
-                    },
-                  ]
-                : []),
-            ]}
-          />
-          <text fg={colors.textMuted}>{lineCount} lines</text>
-        </box>
+        <MarkdownViewer
+          content={content}
+          width={panelWidth}
+          headerLeft={<text fg={colors.primary}>Pentest Report</text>}
+          headerRight={<text fg={colors.textMuted}>{reportPath}</text>}
+          footerLeft={
+            <DialogControls
+              controls={[
+                ...(onOpenExternal
+                  ? [
+                      {
+                        key: "E",
+                        label: "Open in Editor",
+                        variant: "primary" as const,
+                      },
+                    ]
+                  : []),
+              ]}
+            />
+          }
+          footerRight={<text fg={colors.textMuted}>{lineCount} lines</text>}
+        />
       </box>
     </box>
   );
